requirements.txt
python-telegram-bot==21.10
langchain==0.3.27
langchain-ollama==0.3.3
langchain-openai==0.3.19
openai==1.77.0
python-dotenv==1.0.1